--- src/main/java/com/swifty/bank/server/core/domain/account/exception/NoSuchUnitedAccountByUuidException.java
@@ -1,9 +1,0 @@
-package com.swifty.bank.server.core.domain.account.exception;
-
-import java.util.NoSuchElementException;
-
-public class NoSuchUnitedAccountByUuidException extends NoSuchElementException {
-    public NoSuchUnitedAccountByUuidException(String msg) {
-        super(msg);
-    }
-}
