+++ src/main/java/com/swifty/bank/server/exception/account/RequestorAndOwnerOfUnitedAccountIsDifferentException.java
@@ -1,4 +1,4 @@
-package com.swifty.bank.server.core.domain.account.exception;
+package com.swifty.bank.server.exception.account;
 
 public class RequestorAndOwnerOfUnitedAccountIsDifferentException extends IllegalArgumentException {
     public RequestorAndOwnerOfUnitedAccountIsDifferentException(String msg) {
