+++ src/main/java/com/swifty/bank/server/exception/authentication/AuthenticationException.java
@@ -1,4 +1,4 @@
-package com.swifty.bank.server.exception;
+package com.swifty.bank.server.exception.authentication;
 
 public class AuthenticationException extends IllegalArgumentException {
     public AuthenticationException(String msg) {
