+++ src/main/java/com/swifty/bank/server/exception/authentication/NoSuchAuthByUuidException.java
@@ -1,4 +1,4 @@
-package com.swifty.bank.server.exception;
+package com.swifty.bank.server.exception.authentication;
 
 public class NoSuchAuthByUuidException extends AuthenticationException {
     public NoSuchAuthByUuidException(String msg) {
