+++ src/main/java/com/swifty/bank/server/exception/authentication/NotLoggedInCustomerException.java
@@ -1,4 +1,4 @@
-package com.swifty.bank.server.exception;
+package com.swifty.bank.server.exception.authentication;
 
 public class NotLoggedInCustomerException extends AuthenticationException {
     public NotLoggedInCustomerException(String msg) {
