+++ src/main/java/com/swifty/bank/server/exception/authentication/TokenContentNotValidException.java
@@ -1,4 +1,4 @@
-package com.swifty.bank.server.exception;
+package com.swifty.bank.server.exception.authentication;
 
 public class TokenContentNotValidException extends AuthenticationException {
     public TokenContentNotValidException(String msg) {
