+++ src/main/java/com/swifty/bank/server/exception/authentication/TokenExpiredException.java
@@ -1,4 +1,4 @@
-package com.swifty.bank.server.exception;
+package com.swifty.bank.server.exception.authentication;
 
 public class TokenExpiredException extends AuthenticationException {
     public TokenExpiredException(String msg) {
