+++ src/main/java/com/swifty/bank/server/exception/authentication/handler/NotLoggedInExceptionHandler.java
@@ -1,8 +1,8 @@
-package com.swifty.bank.server.exception.handler;
+package com.swifty.bank.server.exception.authentication.handler;
 
 import com.swifty.bank.server.api.service.dto.ResponseResult;
 import com.swifty.bank.server.api.service.dto.Result;
-import com.swifty.bank.server.exception.NotLoggedInCustomerException;
+import com.swifty.bank.server.exception.authentication.NotLoggedInCustomerException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
